chore(cli): Check if klave.json file exists only on specific commands
